refactor(breakout-room): created a reusable `useDragAndDrop` hook in to avoid duplicated code
